"Updated trufflehog version in trunk.yaml, chamged API/app.py file to app.py and reformatted README.md file"
