+++ README.md
@@ -6,7 +6,7 @@
 
 ## Overview
 
-ByteBuilders is a web application designed to showcase our tech business. The project includes a frontend developed with HTML, CSS and JS and a backend using Flask in Python.
+#### ByteBuilders is a web application designed to showcase our tech business. The project includes a frontend developed with HTML, CSS and JS and a backend using Flask in Python.
 
 ## Table of Contents
 
@@ -19,7 +19,7 @@ ByteBuilders is a web application designed to showcase our tech business. The pr
 
 ## Installation
 
-To get started with the ByteBuilders project, follow the instructions below.
+### To get started with the ByteBuilders project, follow the instructions below.
 
 ## Prerequisites
 
@@ -29,77 +29,85 @@ To get started with the ByteBuilders project, follow the instructions below.
 
 ## Setting Up the Project
 
-1. Clone the repository:
+### 1. Clone the repository:
 
-   ```sh
+```
    git clone https://github.com/sogayar/bytebuilders.git
+```
+```
    cd bytebuilders
-
+```
 
 
 ## Usage
-
-2. To use this aplication
-    1. Run the Flask application:
+### 2. To use this aplication
+  #### 1. Run the Flask application:
 ``` 
 python app.py 
 ```
-    2. Open your browser and navigate to http://127.0.0.1:5000/ to view the site.
+  #### 2. Open your browser and navigate to http://127.0.0.1:5000/ to view the site.
 
 
 
 ## Project Structure
 
-3. 
+### 3. The Project Structure
 ```
 ByteBuilders/
-    │
-    │
-    │
-    │── api/ --> Henrique
-    │    │── app.py ### Não está funcionando esta merda
-    │    │── config.py 
-    │    └── models.py 
-    │
-    │
-    │── templates/ --> Caetano
-    │    │── index.html 
-    │    ├── login.html ### Necessário Criar
-    │    └── perfil.html ### Necessário Criar
-    │
-    │── static/ --> Caetano
-    │    │── css/ 
-    │    │    └── styles.css
-    │    │    └── carrocel_styles.css
-    │    │    └── *OUTROS ARQUIVOS CSS*
-    │    │── js/ 
-    │    │    └── script.js
-    │    │    └── carrocel_styles.css
-    │    │    └── *OUTROS ARQUIVOS JS*
-    │    └── img/ 
-    │         └── Imagens
-    │         └── Imagens_Carrossel
-    │
-    │
-    └── README.md
+      │
+      │
+      │
+      │── api/ --> Henrique
+      │    │── app.py ### Não está funcionando esta merda
+      │    │── config.py 
+      │    └── models.py 
+      │
+      │── static/ --> Caetano
+      │    │── css/ 
+      │    │    └── styles.css
+      │    │    └── carrocel_styles.css
+      │    │    └── *OUTROS ARQUIVOS CSS*
+      │    │── js/ 
+      │    │    └── script.js
+      │    │    └── carrocel_styles.css
+      │    │    └── *OUTROS ARQUIVOS JS*
+      │    └── img/ 
+      │         └── Imagens
+      │         └── Imagens_Carrossel
+      │
+      │── templates/ --> Caetano
+      │    │── index.html 
+      │    ├── login.html ### Necessário Criar
+      │    └── perfil.html ### Necessário Criar
+      │
+      │
+      └── README.md
 ```
 
 
 ## Routes
-4.
-    For now, there is only this pages
+### 4. This is the routes that our aplication have
 
 - /: Home page
 - /login: Login page
 - /<nome_usuario>: User profile page
 
 ## Contributing
-5.
+### 5. You can help this projects too
 Contributions are welcome! Please fork the repository and use a feature branch. Pull requests are warmly welcome.
 
 Fork the repository.
-Create a new branch: git checkout -b feature-name
+Create a new branch: 
+```
+git checkout -b feature-name
+```
 Make your changes.
-Commit your changes: git commit -m 'Add some feature'
-Push to the branch: git push origin feature-name
-Submit a pull request.
+Commit your changes: 
+```
+git commit -m 'Add some feature'
+```
+Push to the branch: 
+```
+git push origin feature-name
+```
+Submit a pull request!!
